created base notification app for hub as view, although each notification will create in their respective app
