+++ all_notifications/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AllNotificationsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'all_notifications'
