+++ message_notification/views.py
@@ -4,18 +4,7 @@ from message_notification.models import MessageNotification
 from accounts.models import MyUser
 
 
-def notification_view(request):
-    target_user = request.user
-    new_notifications = []
-    notifications = MessageNotification.objects.all().order_by('-id')
-
-    if len(notifications) > 1:
-        no_new_notifs = 'You have no new notifications!'
-
-    context = {'request': request, 'notifications': notifications, 'no_new_notifs': no_new_notifs}
-    return render(request, 'notifications.html', context)
-
-def create_notification(message, tagged):
+def create_message_notification(message, tagged):
     all_users = MyUser.objects.all()
     names = [x.username for x in all_users]
     print(names)
